requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
anyio==4.12.1
attrs==25.4.0
azure-core==1.37.0
azure-identity==1.25.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
dateparser==1.2.2
dotenv==0.9.9
frozenlist==1.8.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hvac==2.4.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
microsoft-kiota-abstractions==1.9.8
microsoft-kiota-authentication-azure==1.9.8
microsoft-kiota-http==1.9.8
microsoft-kiota-serialization-form==1.9.8
microsoft-kiota-serialization-json==1.9.8
microsoft-kiota-serialization-multipart==1.9.8
microsoft-kiota-serialization-text==1.9.8
msal==1.34.0
msal-extensions==1.3.1
msgraph-core==1.3.8
msgraph-sdk==1.52.0
multidict==6.7.0
numpy==2.4.1
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
pandas==2.3.3
prometheus-api-client==0.7.0
propcache==0.4.1
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
regex==2025.11.3
requests==2.32.5
six==1.17.0
std-uritemplate==2.0.8
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
yarl==1.22.0
zipp==3.23.0
